refactor: Переместил пересечления в папку type
