+++ src/main/java/ru/yandex/practicum/filmorate/model/type/Rating.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.model;
+package ru.yandex.practicum.filmorate.model.type;
 
 public enum Rating {
